Refactors to MustReads and NavMini
